requirements.txt
streamlit
langchain
langchain-openai
langchain-core
langchain-community
python-dotenv
langchain-groq